corrected some simple warning and avoided collapse of edges of triangular faces for polygons

+++ vcg/container/simple_temporary_data.h
@@ -82,7 +82,7 @@ public:
     void resize(size_t sz)
     {
         int oldDatasize = datasize;
-        if (sz <= oldDatasize)
+        if ((int)sz <= oldDatasize)
             return;
         if (sz > datareserve)
             reserve(sz);
